+++ src/Sharpen.Engine/SharpenSuggestions/CSharp50/AsyncAwait/AwaitEquivalentAsynchronousMethod.cs
@@ -1,0 +1,13 @@
+﻿using static Sharpen.Engine.SharpenSuggestions.CSharp50.AsyncAwait.EquivalentAsynchronousMethodFinder;
+
+namespace Sharpen.Engine.SharpenSuggestions.CSharp50.AsyncAwait
+{
+    internal sealed class AwaitEquivalentAsynchronousMethod : BaseAwaitingEquivalentAsynchronousMethod
+    {
+        private AwaitEquivalentAsynchronousMethod() :base(EnclosingMethodAsyncStatus.EnclosingMethodMustBeAsync) { }
+
+        public override string FriendlyName { get; } = "Await equivalent asynchronous method";
+
+        public static readonly AwaitEquivalentAsynchronousMethod Instance = new AwaitEquivalentAsynchronousMethod();
+    }
+}